We now have 4 distinct CA, and a Root signing/managing them
After some strugles, it seems to be the best way to get what we need.
There are more information on Github:
https://github.com/xtreemfs/xtreemfs/issues/349#issuecomment-65408404

The create-cert script will create the four certificates signed by the
four CA. Puppet will then take the right one in order to push it to the
pkcs12 file.